load_wfdb_signals: fix handling of multi-frequency input.
For multi-frequency variable-layout input records, different signals
may have different numbers of samples per frame, and the signal
indices may vary from one segment to the next.  Here, we want to use
the number of samples per frame for signal i in the current segment -
not for signal i in the layout header.