Pin the soundfile package to a hopefully-working version.
There seems to be some problem with the soundfile 0.12.0 and 0.12.1
wheel packages for x86-64 Linux - I haven't completely figured out the
cause of the problem.  In some circumstances, when *reading* a FLAC
signal file with wfdb-python, we get the error "internal psf_fseek()
failed".

The easiest way to avoid this is to use an older soundfile version
which doesn't include Linux binary wheels (but does include binary
wheels for Windows and MacOS.)  On Linux, this will use the system
libsndfile binary that you probably already have installed.