formats/wfdb: implement open/read/close API.
Implement the open_waveforms / close_waveforms / read_opened_waveforms
API as defined in BaseFormat.

Since wfdb-python doesn't have any API for doing this, we implement
I/O for formats 16 and 516 "by hand".  This isn't a complete or
conformant implementation of the WFDB format and not recommended for
applications to use as-is - it should work for the data files produced
by write_waveforms, but not for arbitrary WFDB records.  But it should
be useful in better quantifying the performance overhead of the
wfdb-python API.

For each signal format, a custom reader class is implemented.  Format
16 is implemented using numpy, much like wfdb-python does; this class
could be adapted to handle formats 80, 61, 160, 24, or 32 if needed.

Format 516 is implemented using soundfile, again much like wfdb-python
does (but lacks the workaround for libsndfile bug #431), and could be
adapted to handle formats 508 or 524 if needed.